core/requirements.txt
jsonify
flask
fastapi
uvicorn
sqlalchemy
pytest
psycopg2-binary
httpx
python-dotenv
pytest-asyncio
pytest-anyio
pydantic
redis
openai